tools/requirements-slides.txt
python-pptx==0.6.23
PyYAML==6.0.2
